fix(time-format): share DateWithFormatter and stop reading four-digit years as epoch offsets
Follow-up to #43839. The NaN guard it added lived in two byte-identical
copies of DateWithFormatter, one per table plugin, while the core
fallback in stringifyTimeInput only terminates because of that guard.
Move the class into @superset-ui/core next to the code that depends on
it, point both plugins at the shared export, and pin the invariant with
a core-level test.

While here, narrow the integer-string branch that #38949 introduced.
It treated every digit-only string as epoch milliseconds, so a bare
year like "2017" became two seconds past 1970. A four-digit string is
the ISO 8601 year-only form, which every engine parses as January 1st
of that year, so it now goes through Date parsing. Longer digit strings
keep the epoch handling the pivot table relies on.
