+++ superset-frontend/packages/superset-ui-core/src/time-format/index.ts
@@ -19,6 +19,7 @@
 
 export { default as TimeFormats, LOCAL_PREFIX } from './TimeFormats';
 export { default as TimeFormatter, PREVIEW_TIME } from './TimeFormatter';
+export { default as DateWithFormatter } from './DateWithFormatter';
 export { DEFAULT_D3_TIME_FORMAT } from './D3FormatConfig';
 
 export {
