+++ superset-frontend/packages/superset-ui-core/src/time-format/utils/stringifyTimeInput.ts
@@ -28,8 +28,17 @@ export default function stringifyTimeInput(
   let time: Date;
   if (typeof value === 'string') {
     const trimmed = value.trim();
+    // A bare four-digit string is the ISO 8601 year-only form ("2017"), which
+    // every engine parses as January 1st of that year. Any other integer
+    // string is an epoch timestamp in milliseconds that was stringified on
+    // its way here, e.g. by the pivot table, and is not a valid Date input.
+    const isYear = /^\d{4}$/.test(trimmed);
     const isIntegerString = /^-?\d+$/.test(trimmed);
-    time = new Date(isIntegerString ? Number(trimmed) : value);
+    if (isYear) {
+      time = new Date(trimmed);
+    } else {
+      time = new Date(isIntegerString ? Number(trimmed) : value);
+    }
   } else {
     time = value instanceof Date ? value : new Date(value);
   }
@@ -37,7 +46,9 @@ export default function stringifyTimeInput(
   // An input that does not resolve to a valid date - a duration such as
   // "00:01:54", for instance - would otherwise be formatted from an Invalid
   // Date and render as "NaN:NaN:NaN". Fall back to its own representation,
-  // as is already done for null and undefined above.
+  // as is already done for null and undefined above. For a `DateWithFormatter`
+  // this calls its `toString()`, which returns the original input rather than
+  // re-entering the formatter; that guard is what keeps the fallback finite.
   if (Number.isNaN(time.getTime())) {
     return `${value}`;
   }
