+++ superset-frontend/packages/superset-ui-core/test/time-format/DateWithFormatter.test.ts
@@ -1,4 +1,4 @@
-/*
+/**
  * Licensed to the Apache Software Foundation (ASF) under one
  * or more contributor license agreements.  See the NOTICE file
  * distributed with this work for additional information
@@ -16,11 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-import { getTimeFormatter } from '@superset-ui/core';
-import { GenericDataType } from '@apache-superset/core/common';
-import DateWithFormatter from '../../src/utils/DateWithFormatter';
-import { formatColumnValue } from '../../src/utils/formatValue';
-import { DataColumnMeta } from '../../src/types';
+import { DateWithFormatter, getTimeFormatter } from '@superset-ui/core';
 
 const formatter = getTimeFormatter('%H:%M:%S');
 
@@ -42,19 +38,4 @@ test('renders the original value when it is not a parseable timestamp', () => {
 test('retains the original input when the formatter is String', () => {
   const value = new DateWithFormatter('00:01:54');
   expect(String(value)).toBe('00:01:54');
-});
-
-test('renders a duration cell through the column formatter without producing NaN', () => {
-  // The cell text is produced by formatColumnValue, which hands the wrapped
-  // value straight to the formatter rather than going through toString().
-  const column: DataColumnMeta = {
-    key: 'call_period',
-    label: 'call_period',
-    dataType: GenericDataType.Temporal,
-    formatter,
-    isNumeric: false,
-  };
-  const value = new DateWithFormatter('00:01:54', { formatter });
-
-  expect(formatColumnValue(column, value)).toEqual([false, '00:01:54']);
 });
